backend/requirements.txt
Flask==3.1.1
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
cohere==5.15.0
eventlet==0.40.0
SQLAlchemy==2.0.41
python-dotenv==1.1.0
gunicorn
